컴포넌트 구조 리팩토링
