+++ next.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  images: {
+    domains: ["firebasestorage.googleapis.com"],
+  },
+};
