+++ pages/game/edit/[gid].tsx
@@ -11,41 +11,14 @@ import {
   Loader,
   Message,
 } from "semantic-ui-react";
-import { useUserContext } from "../../../src/common/contexts/UserContext";
-import styles from "../../../styles/game/edit.module.css";
 import firebase from "firebase/app";
 import "firebase/storage";
 
-function getExt(filename: string) {
-  return filename
-    .substring(filename.lastIndexOf("."), filename.length)
-    .toLowerCase();
-}
+import styles from "../../../styles/game/edit.module.css";
 
-export const checkYoutubeUrl = (gid: string) => {
-  var exptext =
-    /^((?:https?:)?\/\/)?((?:www|m)\.)?((?:youtube\.com|youtu.be))(\/(?:[\w\-]+\?v=|embed\/|v\/)?)([\w\-]+)(\S+)?$/;
-  if (exptext.test(gid) == false) {
-    return false;
-  }
-  return true;
-};
-
-export const checkGid = (gid: string) => {
-  var exptext = /^G\-[A-Z0-9]{3}\-[A-Z0-9]{3}\-[A-Z0-9]{3}$/;
-  if (exptext.test(gid) == false) {
-    return false;
-  }
-  return true;
-};
-
-export const checkPid = (gid: string) => {
-  var exptext = /^P\-[A-Z0-9]{3}\-[A-Z0-9]{3}\-[A-Z0-9]{3}$/;
-  if (exptext.test(gid) == false) {
-    return false;
-  }
-  return true;
-};
+import { useUserContext } from "../../../src/common/contexts/UserContext";
+import { getExt } from "../../../src/common/utils/file";
+import { checkGid, checkPid, checkYoutubeUrl } from "../../../src/common/utils/regex";
 
 const EditGamePage = () => {
   const router = useRouter();
