+++ src/common/utils/file.ts
@@ -1,0 +1,5 @@
+export function getExt(filename: string) {
+  return filename
+    .substring(filename.lastIndexOf("."), filename.length)
+    .toLowerCase();
+}
