+++ src/common/utils/regex.ts
@@ -1,0 +1,24 @@
+export const checkYoutubeUrl = (gid: string) => {
+  var exptext =
+    /^((?:https?:)?\/\/)?((?:www|m)\.)?((?:youtube\.com|youtu.be))(\/(?:[\w\-]+\?v=|embed\/|v\/)?)([\w\-]+)(\S+)?$/;
+  if (exptext.test(gid) == false) {
+    return false;
+  }
+  return true;
+};
+
+export const checkGid = (gid: string) => {
+  var exptext = /^G\-[A-Z0-9]{3}\-[A-Z0-9]{3}\-[A-Z0-9]{3}$/;
+  if (exptext.test(gid) == false) {
+    return false;
+  }
+  return true;
+};
+
+export const checkPid = (gid: string) => {
+  var exptext = /^P\-[A-Z0-9]{3}\-[A-Z0-9]{3}\-[A-Z0-9]{3}$/;
+  if (exptext.test(gid) == false) {
+    return false;
+  }
+  return true;
+};
